[fix] Rename view files
