+++ src/views/details-modal.js
