+++ src/views/movie.js
