+++ src/views/movies.js
